# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Swagger API",
        "description": "My Swagger API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/val": {
            "post": {
                "tags": [
                    "External APIs"
                ],
                "summary": "Exchange currency API",
                "operationId": "exchangeAPI",
                "requestBody": {
                    "description": "Exchange API data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "from_date": {
                                        "description": "Date from in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-15"
                                    },
                                    "to_date": {
                                        "description": "Date to in YYYY-MM-DD format",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-15"
                                    },
                                    "val_code": {
                                        "description": "Currency code.",
                                        "type": "string",
                                        "example": "usd"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of currencies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The request was malformed or invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/email": {
            "post": {
                "tags": [
                    "Email"
                ],
                "summary": "Send Email",
                "operationId": "sendEmail",
                "requestBody": {
                    "description": "Send email data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "to_email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "test_user@test.com"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title email"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Content letter email"
                                    },
                                    "attachments": {
                                        "type": "array",
                                        "items": {
                                            "description": "Array of file attachments to email. Supports file types: pdf, docx, doc, jpg, jpeg, png, gif",
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}